\name{MANHATAN}
\alias{MANHATAN}
\docType{data}

\title{The Midtown Manhattan Study}

\description{
The \code{MANHATAN} data frame has 25 rows and 3 columns.
The observations were obtained in a study carried out by the sociologist Leo Srole and describe the cross-classification of 1660 adults in Manhattan, ages 20-59, obtained from a sample of midtown residents.
}

\format{
 This data frame contains the following columns:
 \describe{
  \item{health}{
   A factor with levels:
   \code{Well};
   \code{Misy};
   \code{Mosy};
   \code{Imp}.
  }
  \item{socecon}{
   A factor with levels:
   \code{A};
   \code{B};
   \code{C};
   \code{D};
   \code{E};
   \code{F}.
  }
  \item{value}{
   The absolute frequencies of which factor. 
  }
 }
}

\source{
 Goodman, L. A. (1987) New Methods for Analysing the Intrinsic Character of Qualitative Variables Using Cross-Classified Data. \emph{American Journal of Sociology} \bold{93}, 529--583.
}

\references{
 van der Ark, A. L. 1999. \emph{Contributions to Latent Budget Analysis, a tool for the analysis of compositional data}. Ph.D. Thesis University of Utrecht. 
}

\usage{
MANHATAN
}

\keyword{datasets}