\name{PerfMark}
\alias{PerfMark}
\docType{data}

\title{BEAUTY SALON MANAGEMENT}

\description{
The \code{PerfMark} data frame has 31 rows and 46 columns.
The data set is the result of a survey of 47 beauty salons located at the city of Lavras, Brazil, consisting of two types of questions; the first identifies the profile of the owner manager (explanatory variable), the second are questions referring to the degree of professionalism with respect to planing, market and finances (response variable).
The data set is already cross-tabulated.
}

\format{
 This data frame contains the following columns referring the absolute frequencies to each row variable:
 \describe{
  Planning variables:
  \item{PA14}{What is the dependence of the owner to function properly?.}
  \item{PA20}{What are your plans towards next year? only a dream.}
  \item{PA21}{What are your plans towards next year? vague goals.}
  Marketing variables:
  \item{MA11}{Your business tries to systematically assess the customer satisfaction and use that as a basis for management decisions. Alternative 1.}
  \item{MA12}{Your business tries to systematically assess the customer satisfaction and use that as a basis for management decisions. Alternative 2.}
  \item{MA20}{Your business offers more than the usual services. Alternative 0.}
  \item{MA21}{Your business offers more than the usual services. Alternative 1.}
  \item{MA30}{Your business is focused to further customer loyalty. Alternative 0.}
  \item{MA31}{Your business is focused to further customer loyalty. Alternative 1.}
  \item{MA32}{Your business is focused to further customer loyalty. Alternative 2.}
  \item{MA42}{What is the proportion, among current customers, of those who are customers for more than 6 months. Alternative 2.}
  \item{MA43}{What is the proportion, among current customers, of those who are customers for more than 6 months. Alternative 3.}
  \item{MB12}{Your business offers more services than when it began. Alternative 2.}
  \item{MB22}{How is your business quality perceived as compared to the competition? Alternative 2.}
  \item{MB23}{How is your business quality perceived as compared to the competition? Alternative 3.}
  \item{MB31}{How is your business range of services perceived as compared to the competition? Alternative 1.}
  \item{MB32}{How is your business range of services perceived as compared to the competition? Alternative 2.}
  \item{MC11}{What is your business level of prices perceived as compared to the competition? Alternative 1.}
  \item{MC12}{What is your business level of prices perceived as compared to the competition? Alternative 2.}
  \item{MD13}{Your business location is perceived as appropriate to the target market. Alternative 3.}
  \item{ME10}{Your business uses formal media to advertise itself. Alternative 0.}
  \item{ME11}{Your business uses formal media to advertise itself. Alternative 1.}
  \item{ME25}{Your business uses formal media to advertise itself. Alternative 5.}
  Financial variables:
  \item{F10 }{Your business clearly separates the owner bills from the business bills. Alternative 0.}
  \item{F14 }{Your business clearly separates the owner bills from the business bills. Alternative 4.}
  \item{F20 }{Your owners withdrawal are planned and controlled in advance. Alternative 0.}
  \item{F21 }{Your owners withdrawal are planned and controlled in advance. Alternative 1.}
  \item{F24 }{Your owners withdrawal are planned and controlled in advance. Alternative 4.}
  \item{F31 }{Your business pays for its purchases in installments. Alternative 1.}
  \item{F34 }{Your business pays for its purchases in installments. Alternative 4.}
  \item{F42 }{Your business knows today whether it will be able to pay its short-term bills of 60 days. Alternative 2.}
  \item{F44 }{Your business knows today whether it will be able to pay its short-term bills of 60 days. Alternative 4.}
  \item{F50 }{Your business uses short-term cash-flow analysis to plan for its short-term bills. Alternative 0.}
  \item{F51 }{Your business uses short-term cash-flow analysis to plan for its short-term bills. Alternative 1.}
  \item{F63 }{Your business has formal control of the monthly amount it makes from its services. Alternative 3.}
  \item{F64 }{Your business has formal control of the monthly amount it makes from its services. Alternative 4.}
  \item{F70 }{Your business uses either credit card, checkbook payment or loans, to finance its needs for working capital. Alternative 0.}
  \item{F74 }{Your business uses either credit card, checkbook payment or loans, to finance its needs for working capital. Alternative 4.}
  \item{F80 }{Your business uses specific credit to finance its needs for capital. Alternative 0.}
  \item{F91 }{The company demonstrates knowledge to properly assess the costs of products used in services and costs of renting and taxes. Alternative 1.}
  \item{F93 }{The company demonstrates knowledge to properly assess the costs of products used in services and costs of renting and taxes. Alternative 3.}
  \item{F100}{Your business clearly identifies the need for working capital. Alternative 0.}
  \item{F111}{Your business lays down the price of services in a systematic way. Alternative 1.}
  \item{F113}{Your business lays down the price of services in a systematic way. Alternative 3.}
  \item{F120}{The company calculates the interest on contracted loans. Alternative 0.}
  \item{F125}{The company calculates the interest on contracted loans. Alternative 5.}
 }
}

\source{
 Jelihovschi, E.G., Alves, R.R., and Correa, F.M. 2011. \emph{Interacting latent budget analysis and correspondence analysis to analyze beauty salon management data}. Biometric Brazilian Journal, 29, 657-673.
}

\references{
 van der Ark, A. L. 1999. \emph{Contributions to Latent Budget Analysis, a tool for the analysis of compositional data}. Ph.D. Thesis University of Utrecht. 
}

\usage{
data(PerfMark)
}

\keyword{datasets}
