\name{pregnancy}
\alias{pregnancy}
\docType{data}

\title{Pregnancy-Related Mortality in California}

\description{
The \code{pregnancy} matrix has 16 rows and 5 columns. The raw data refers to California pregnancy-related deaths from 2002-2005.
}

\format{
 This matrix contains the following columns:
 \describe{
  \item{Pre.E}{bla}
  \item{OH}{bla}
  \item{CVD}{bla}
  \item{DVTPE}{bla}
  \item{AFE}{bla}
 }
 The rows refers to:
 \describe{
  \item{Hfob}{bla}
  \item{Husb}{bla}
  \item{Wnh}{bla}
  \item{Bnh}{bla}
  \item{$<30$b}{bla}
  \item{$30-40$b}{bla}
  \item{$>40$b}{bla}
  \item{$1$}{bla}
  \item{$2-4$}{bla}
  \item{$5+$}{bla}
  \item{$<30$a}{bla}
  \item{$30-40$a}{bla}
  \item{$>40$a}{bla}
  \item{$<32$w}{bla}
  \item{$32-36$w}{bla}
  \item{$>37$w}{bla}
 }
}

\source{
Main, E. K.; et al. Pregnancy-Related Mortality in California: Causes, Characteristics, and Improvement Opportunities. \emph{OBSTETRICS \& GYNECOLOGY}. \bold{125}, 938--947.
}

\references{
Main, E. K.; et al. Pregnancy-Related Mortality in California: Causes, Characteristics, and Improvement Opportunities. \emph{OBSTETRICS \& GYNECOLOGY}. \bold{125}, 938--947.
}

\usage{
pregnancy
}

\keyword{dataset}